requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cloudinary==1.8.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
pbr==3.1.1
Pillow==5.1.0
pycurl==7.43.0
pytesseract==0.2.0
requests==2.18.4
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
xlrd==1.1.0
